backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
boto3==1.34.0
python-multipart==0.0.6
PyJWT==2.9.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
email-validator==2.1.0.post1
